Improve view for Active-Standby Brokers
When the monitor is connected to a broker that is in the Standby role, endpoint information can be misleading. e.g. Bind count is zero, yet that is expected with all binds on the Active role broker.

A clear message is displayed informing of this instead of showing endpoint information.